requirements.txt
aiofiles==24.1.0
aiogram==2.2
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==4.2.17
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
magic-filter==1.0.12
multidict==6.1.0
pillow==11.0.0
propcache==0.2.1
pydantic==2.9.2
pydantic-core==2.23.4
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.18.3
